% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcsp.R
\name{print.dbcsp}
\alias{print.dbcsp}
\title{Print function implemented by dbcsp class}
\usage{
\method{print}{dbcsp}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[=dbcsp-class]{dbcsp}}.}

\item{...}{not currently used.}
}
\value{
No return value, called for side effects.
}
\description{
This function prints information about \code{\link[=dbcsp-class]{dbcsp}} class.
}
\details{
It provides information about the object and the class.
}
\examples{
# Read data from 2 classes
x <- AR.data$come[1:30]
y <- AR.data$five[1:30]
mydbcsp <- new("dbcsp", X1 = x, X2 = y)
print(mydbcsp)
}
\seealso{
\code{\link[=dbcsp-class]{dbcsp}}, \code{\link[=summary.dbcsp]{summary}}, \code{\link[=train.dbcsp]{train}}, \code{\link{selectQ}}, \code{\link[=predict.dbcsp]{predict}}, \code{\link[=plot.dbcsp]{plot}}, \code{\link[=boxplot.dbcsp]{boxplot}}
}
